requirements.txt

requests
redis
fpdf2
fastapi
uvicorn
pytz
